requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
gunicorn==22.0.0
